docker-deploy/web-app/requirements.txt
asgiref==3.2.3
Django==3.0.2
django-crispy-forms==1.8.1
django-flatpickr==1.0.1
psycopg2
pytz==2019.3
sqlparse==0.3.0